Fix for HARMONY-2084 [drlvm][jit][opt] Jitrino.OPT inliner can cause native SOE exception for some large methods + regression test
Fix for HARMONY-4708 [drlvm][jit][opt] Assorted bcmapping and logging related improvements in Jitrino.OPT compiler



git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/drlvm/trunk@571484 13f79535-47bb-0310-9956-ffa450edef68

+++ vm/jitrino/src/codegenerator/ia32/Ia32InstCodeSelector.cpp
@@ -129,7 +129,7 @@ float   __stdcall   convI8F4    (uint64 v) {    return (float)(int64)v; }
 // FP remainder internal helpers (temp solution to be optimized)
 float   __stdcall   remF4   (float v0, float v1)stdcall__;
 float   __stdcall   remF4   (float v0, float v1)   { 
-    return fmod((double)v0,(double)v1);
+    return (float)fmod((double)v0,(double)v1);
 }
 
 double  __stdcall   remF8   (double v0, double v1)stdcall__;
